backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.1
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-multipart>=0.0.9
google-api-python-client>=2.116.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
pandas>=2.2.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
